+++ R/globals.R
@@ -1,1 +1,0 @@
-.globals <- new.env(hash = TRUE)