+++ NEWS.md
@@ -18,6 +18,7 @@ No longer force render data as JSON if using an HTML template.
 - Added `get_header`, `set_header`, and `set_headers` to `Response`.
 - Allow adding multiple cookies to the request.
 - Add `parseCookie` JavaScript helper function to JavaScript file.
+- Allow customising the cookie parser with `as_cookie_parser`.
 
 # ambiorix 2.0.0